conform to tsdk 2.0.0-alpha17 / csdk 2.0.0-alpha18

+++ lib/ZitiMfaAuthQuery.swift
@@ -34,10 +34,10 @@ import CZitiPrivate
     public var httpUrl:String?
     
     /// minimum length
-    public var minLength:Int32?
+    public var minLength:Int64?
     
     /// maximum length
-    public var maxLength:Int32?
+    public var maxLength:Int64?
     
     /// expected format
     public var format:String?
